// src/app/models/grower-account.interface.ts
export interface IGrowerAccount {
	id: number;
	growerId: number;
	accountType: string;
	accountSuffix: string;
	dateOpened: Date | string | null;
	deductChargeInterestFromDate: Date | string | null;
	interestRate: number | null;
	noFlocksToPay: number | null;
	remainingFlocksToPay: number | null;
	balanceGoalStartDeduction: number | null;
	beginYearBalance: number | null;
	balanceForward: number | null;
	currentCharges: number | null;
	lastChargeDate: Date | string | null;
	currentCredits: number | null;
	lastCreditDate: Date | string | null;
	cashAdvances: number | null;
	amountDue: number | null;
	ytdinterestCharged: number | null;
	ytdinterestPaid: number | null;
	accountComment: string;
}